窗帘控制器添加场景中，选择设备时选择其设置为开启或关闭；添加单路插座；窗帘控制器的MCU升级；其他对RGB设备的操作终止录音和循环；插座清除电量统计数据；插座功率异常门限值

+++ CSRmeshDemo/Main/Controller/MusicListTableViewController.m
@@ -98,7 +98,7 @@
     }else {
         cell.imageView.image = nil;
         MPMediaItem *item = self.dataArray[indexPath.row];
-        cell.textLabel.text = [NSString stringWithFormat:@"%ld   %@",indexPath.row,[item valueForProperty:MPMediaItemPropertyTitle]];
+        cell.textLabel.text = [NSString stringWithFormat:@"%ld   %@",(long)indexPath.row,[item valueForProperty:MPMediaItemPropertyTitle]];
     }
     
     return cell;
